+++ frontends/web/react/components/index.ts
@@ -1,1 +1,5 @@
+export * from "./audio-gb/audio-gb";
+export * from "./debug/debug";
 export * from "./help/help";
+export * from "./registers-gb/registers-gb";
+export * from "./tiles-gb/tiles-gb";