+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * Support for variable clock speed for APU, means variable audio speed
+* Moved debug into the base emulator (from emukit)
 
 ### Changed